rename all async, changed package json to only build js files, moved .h files to incude directory
