+++ tools/nodejs_api/src_cpp/include/node_query_result.h
