+++ tools/nodejs_api/src_cpp/include/util.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <napi.h>
+#include "main/kuzu.h"
+
+class Util {
+    public:
+        static Napi::Value ConvertToNapiObject(const kuzu::common::Value& value, Napi::Env env);
+};
